Migrate PostCover to StaticQuery.
