+++ src/components/PostCover/PostCover.jsx
@@ -1,50 +1,59 @@
 import React, { Component } from "react";
-import Img from "gatsby-image";
-import path from "path";
-import "./PostCover.scss";
+import { StaticQuery, graphql } from "gatsby";
+import PostCover from "./PostCoverComponent";
 
-class PostCover extends Component {
+class queryWrapper extends Component {
   render() {
-    const { fileEdges, postNode, coverHeight, coverClassName } = this.props;
-    const post = postNode.frontmatter ? postNode.frontmatter : postNode;
-    const coverNodeList = fileEdges.filter(fileNode => {
-      if (fileNode.node.childImageSharp === null) return false;
-
-      if (
-        fileNode.node.absolutePath.indexOf(
-          path.join("/static/assets/", post.cover)
-        ) !== -1
-      )
-        return true;
-
-      return false;
-    });
-
-    if (coverNodeList.length === 1) {
-      return (
-        <Img
-          sizes={coverNodeList[0].node.childImageSharp.sizes}
-          outerWrapperClassName={coverClassName}
-          style={{ height: coverHeight, width: "100%" }}
-        />
-      );
-    }
-
-    /* eslint no-undef: "off" */
-    const coverURL =
-      post.cover.substring(0, 1) === "/"
-        ? __PATH_PREFIX__ + post.cover
-        : post.cover;
+    const { postNode, coverHeight, coverClassName } = this.props;
     return (
-      <div
-        style={{
-          backgroundImage: `url(${coverURL})`,
-          height: `${coverHeight}px`
-        }}
-        className={coverClassName}
+      <StaticQuery
+        query={graphql`
+          query CoverQuery {
+            allFile {
+              edges {
+                node {
+                  id
+                  absolutePath
+                  childImageSharp {
+                    id
+                    resolutions {
+                      base64
+                      tracedSVG
+                      aspectRatio
+                      width
+                      height
+                      src
+                      srcSet
+                      srcWebp
+                      srcSetWebp
+                      originalName
+                    }
+                    internal {
+                      contentDigest
+                      type
+                      owner
+                    }
+                    sizes(maxWidth: 1240) {
+                      ...GatsbyImageSharpSizes
+                      originalName
+                    }
+                  }
+                }
+              }
+            }
+          }
+        `}
+        render={data => (
+          <PostCover
+            fileEdges={data.allFile.edges}
+            postNode={postNode}
+            coverHeight={coverHeight}
+            coverClassName={coverClassName}
+          />
+        )}
       />
     );
   }
 }
 
-export default PostCover;
+export default queryWrapper;
