+++ src/components/PostListing/PostListing.jsx
@@ -23,11 +23,7 @@ class PostListing extends React.Component {
       <div className="md-grid md-grid--no-spacing md-cell--middle">
         <div className="md-grid md-cell--8 mobile-fix">
           {postList.map(post => (
-            <PostPreview
-              key={post.title}
-              postInfo={post}
-              fileEdges={this.props.fileEdges}
-            />
+            <PostPreview key={post.title} postInfo={post} />
           ))}
         </div>
       </div>
