+++ src/pages/index.jsx
@@ -9,7 +9,6 @@ import config from "../../data/SiteConfig";
 class Index extends React.Component {
   render() {
     const postEdges = this.props.data.allMarkdownRemark.edges;
-    const fileEdges = this.props.data.allFile.edges;
     return (
       <Layout location={this.props.location}>
         <div className="index-container">
@@ -18,7 +17,7 @@ class Index extends React.Component {
             <link rel="canonical" href={`${config.siteUrl}`} />
           </Helmet>
           <SEO postEdges={postEdges} />
-          <PostListing postEdges={postEdges} fileEdges={fileEdges} />
+          <PostListing postEdges={postEdges} />
         </div>
       </Layout>
     );
@@ -29,38 +28,6 @@ export default Index;
 
 export const pageQuery = graphql`
   query IndexQuery {
-    allFile {
-      edges {
-        node {
-          id
-          absolutePath
-          childImageSharp {
-            id
-            resolutions {
-              base64
-              tracedSVG
-              aspectRatio
-              width
-              height
-              src
-              srcSet
-              srcWebp
-              srcSetWebp
-              originalName
-            }
-            internal {
-              contentDigest
-              type
-              owner
-            }
-            sizes(maxWidth: 1240) {
-              ...GatsbyImageSharpSizes
-              originalName
-            }
-          }
-        }
-      }
-    }
     allMarkdownRemark(
       limit: 2000
       sort: { fields: [fields___date], order: DESC }
