+++ src/templates/category.jsx
@@ -9,7 +9,6 @@ export default class CategoryTemplate extends React.Component {
   render() {
     const { category } = this.props.pageContext;
     const postEdges = this.props.data.allMarkdownRemark.edges;
-    const fileEdges = this.props.data.allFile.edges;
     return (
       <Layout location={this.props.location}>
         <div className="category-container">
@@ -22,7 +21,7 @@ export default class CategoryTemplate extends React.Component {
               href={`${config.siteUrl}/categories/${category}`}
             />
           </Helmet>
-          <PostListing postEdges={postEdges} fileEdges={fileEdges} />
+          <PostListing postEdges={postEdges} />
         </div>
       </Layout>
     );
@@ -31,38 +30,6 @@ export default class CategoryTemplate extends React.Component {
 
 export const pageQuery = graphql`
   query CategoryPage($category: String) {
-    allFile {
-      edges {
-        node {
-          id
-          absolutePath
-          childImageSharp {
-            id
-            resolutions {
-              base64
-              tracedSVG
-              aspectRatio
-              width
-              height
-              src
-              srcSet
-              srcWebp
-              srcSetWebp
-              originalName
-            }
-            internal {
-              contentDigest
-              type
-              owner
-            }
-            sizes(maxWidth: 1240) {
-              ...GatsbyImageSharpSizes
-              originalName
-            }
-          }
-        }
-      }
-    }
     allMarkdownRemark(
       limit: 1000
       sort: { fields: [fields___date], order: DESC }
