+++ src/templates/tag.jsx
@@ -9,7 +9,7 @@ export default class TagTemplate extends React.Component {
   render() {
     const { tag } = this.props.pageContext;
     const postEdges = this.props.data.allMarkdownRemark.edges;
-    const fileEdges = this.props.data.allFile.edges;
+
     return (
       <Layout location={this.props.location}>
         <div className="tag-container">
@@ -17,7 +17,7 @@ export default class TagTemplate extends React.Component {
             <title>{`Posts tagged as "${tag}" | ${config.siteTitle}`}</title>
             <link rel="canonical" href={`${config.siteUrl}/tags/${tag}`} />
           </Helmet>
-          <PostListing postEdges={postEdges} fileEdges={fileEdges} />
+          <PostListing postEdges={postEdges} />
         </div>
       </Layout>
     );
@@ -26,38 +26,6 @@ export default class TagTemplate extends React.Component {
 
 export const pageQuery = graphql`
   query TagPage($tag: String) {
-    allFile {
-      edges {
-        node {
-          id
-          absolutePath
-          childImageSharp {
-            id
-            resolutions {
-              base64
-              tracedSVG
-              aspectRatio
-              width
-              height
-              src
-              srcSet
-              srcWebp
-              srcSetWebp
-              originalName
-            }
-            internal {
-              contentDigest
-              type
-              owner
-            }
-            sizes(maxWidth: 1240) {
-              ...GatsbyImageSharpSizes
-              originalName
-            }
-          }
-        }
-      }
-    }
     allMarkdownRemark(
       limit: 1000
       sort: { fields: [fields___date], order: DESC }
